# demo/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ticket Tracker API
  description: API for ticket management
  version: 1.0.0
paths:

  /api/projects:
    get:
      summary: List projects (paginated)
      description: Returns a paginated list of all projects.
      operationId: getProjects
      tags: [ Projects ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 1000
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ProjectPage'

    post:
      summary: Create project
      description: Creates a new project using the provided name and description.
      operationId: createProject
      tags: [ Projects ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ProjectDto'
        '400':
          $ref: 'schemas.yaml#/components/responses/BadRequest'

  /api/projects/{id}:
    get:
      summary: Get project
      description: Retrieves the detailed information of a single project by its ID.
      operationId: getProjectById
      tags: [ Projects ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: 'schemas.yaml#/components/schemas/ProjectDto' }
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

    delete:
      summary: Delete project
      description: Deletes an existing project by its ID.
      operationId: deleteProject
      tags: [ Projects ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

  /api/tickets:
    get:
      summary: List/filter tickets (paginated)
      description: Returns a paginated list of tickets.
      operationId: getTickets
      tags: [ Tickets ]
      parameters:
        - name: status
          in: query
          schema:
            $ref: 'schemas.yaml#/components/schemas/Status'
        - name: priority
          in: query
          schema:
            $ref: 'schemas.yaml#/components/schemas/Priority'
        - name: assigneeId
          in: query
          schema:
            type: string
            format: uuid
        - name: dueBefore
          in: query
          schema:
            type: string
            format: date-time
        - name: dueAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: titleContains
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 1000
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [ dueDate, priority, createdAt ]
        - name: sortDirection
          in: query
          schema:
            type: string
            enum: [ asc, desc ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/TicketPage'

    post:
      summary: Create ticket
      description: Create a new ticket.
      operationId: createTicket
      tags: [ Tickets ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/TicketDto'
        '400':
          $ref: 'schemas.yaml#/components/responses/BadRequest'

  /api/tickets/{id}:
    get:
      summary: Get ticket
      description: Get a single ticket by its ID.
      operationId: getTicketById
      tags: [ Tickets ]
      parameters:
        - $ref: 'schemas.yaml#/components/parameters/TicketId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/TicketDto'
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

    put:
      summary: Full update ticket
      description: Replaces all parameters of an existing ticket by its ID.
      operationId: updateTicket
      tags: [ Tickets ]
      parameters:
        - $ref: 'schemas.yaml#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/TicketPutRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/TicketDto'
        '400':
          $ref: 'schemas.yaml#/components/responses/BadRequest'
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

    patch:
      summary: Partial update ticket
      description: Update one or more parameters (eg. status only)
      operationId: patchTicket
      tags: [ Tickets ]
      parameters:
        - $ref: 'schemas.yaml#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/TicketPatchRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/TicketDto'
        '400':
          $ref: 'schemas.yaml#/components/responses/BadRequest'
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

    delete:
      summary: Delete ticket
      description: Deletes an existing ticket by its ID.
      operationId: deleteTicket
      tags: [ Tickets ]
      parameters:
        - $ref: 'schemas.yaml#/components/parameters/TicketId'
      responses:
        '204':
          description: No Content
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

  /api/tickets/{ticketId}/comments:
    get:
      summary: List comments for a ticket
      description: Retrieves a paginated list of comments associated with a specific ticket ID.
      operationId: getTicketComments
      tags: [ Comments ]
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 1000
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/CommentPage'
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

    post:
      summary: Add comment
      description: Adds a new comment authored by the current user to the specified ticket.
      operationId: createComment
      tags: [ Comments ]
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/CommentDto'
        '400':
          $ref: 'schemas.yaml#/components/responses/BadRequest'
        '404':
          $ref: 'schemas.yaml#/components/responses/NotFound'

  /api/auth/register:
    post:
      summary: Register a new local account
      operationId: register
      tags: [ Auth ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/RegisterResponse'
        '400':
          $ref: 'schemas.yaml#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: Authenticate and get tokens
      operationId: login
      tags: [ Auth ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/AuthResponse'
        '400':
          $ref: 'schemas.yaml#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ErrorResponse'

  /api/auth/refresh:
    post:
      summary: Exchange refresh token for a new pair
      operationId: refresh
      tags: [ Auth ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/AuthResponse'
        '401':
          description: Invalid, expired or revoked refresh token
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      summary: Revoke a specific refresh token
      operationId: logout
      tags: [ Auth ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: No Content

  /api/auth/logout-all:
    post:
      summary: Revoke all refresh tokens for the current user
      operationId: logoutAll
      tags: [ Auth ]
      security:
        - bearerAuth: [ ]
      responses:
        '204':
          description: No Content

  /api/auth/google:
    get:
      summary: Start Google OAuth login
      description: |
        PASTE THIS ON BROWSER: `http://localhost:8080/api/auth/google` 
        
        Redirects the browser to Google's authorization page.
        After the user authenticates, Google redirects back to
        /api/login/oauth2/code/google (handled by Spring Security).
      operationId: googleLogin
      tags: [ Auth ]
      responses:
        '302':
          description: Redirect to Google authorization endpoint
          headers:
            Location:
              schema:
                type: string
                example: /oauth2/authorization/google